requirements.txt
aiogram>=3.3.0,<4.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
aiohttp>=3.9.0
httpx>=0.25.0
apscheduler>=3.10.0
pillow>=10.0.0
groq>=0.4.0
google-generativeai>=0.3.0
